pyproject.toml
[project]
name = "atl-transit-data"
version = "0.1.0"
description = "Ingesting and Transforming Atlanta Transit Data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-bigquery==1.9.1",
    "google-cloud-storage>=2.19.0",
    "gtfs-realtime-bindings>=1.0.0",
    "requests>=2.32.3",
]
